Add filters to logging, update docs
